+++ CHANGELOG.md
@@ -14,12 +14,16 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - mirrord-config: Fix disabled feature for env in config file, `env = false` should work. See [#1015](https://github.com/metalbear-co/mirrord/issues/1015).
 - VS Code extension: release universal extension as a fallback for Windows and other platforms to be used with WSL/Remote development. Fixes [#1017](https://github.com/metalbear-co/mirrord/issues/1017)
 - Fix `MIRRORD_AGENT_RUST_LOG` can't be more than info due to dependency on info log.
+- Fix pause feature not working in extension due to writing to stdout (changed to use trace)
 
 ### Changed
 
 - `DNSLookup` failures changed to be info log from error since it is a common case.
 - mirrord-agent: now prints "agent ready" instead of logging it so it can't be fudged with `RUST_LOG` control.
 - mirrord-agent: `agent::layer_recv` changed instrumentation to be trace instead of info.
+- mirrord-layer/agent: change ttl of job to be 1 second for cases where 0 means in cluster don't clean up.
+- Convert go fileops e2e tests into integration tests. Part of
+  [#994](https://github.com/metalbear-co/mirrord/issues/994#issuecomment-1410721960).
 
 ## 3.22.0